# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QPX Quantum Progress Experience API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /player/profile:
    get:
      operationId: getPlayerProfile
      tags: [player]
      summary: Get player profile with RPG stats
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePlayerProfile
      tags: [player]
      summary: Update player profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
  /player/onboarding:
    post:
      operationId: completeOnboarding
      tags: [player]
      summary: Complete onboarding questionnaire and create player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
  /player/xp:
    post:
      operationId: awardXp
      tags: [player]
      summary: Award XP to player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XpAwardInput"
      responses:
        "200":
          description: Updated player with XP result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XpAwardResult"

  # ── Quests ──────────────────────────────────────────────────────────────────
...
  /journal:
    get:
      operationId: getJournalEntries
      tags: [journal]
      summary: List journal entries
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Created journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journal/{id}:
    get:
      operationId: getJournalEntry
      tags: [journal]
      summary: Get a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryUpdate"
      responses:
        "200":
          description: Updated journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Stats ────────────────────────────────────────────────────────────────────
...
  /boss-battles/{id}/progress:
    patch:
      operationId: updateBossBattleProgress
      tags: [bossBattles]
      summary: Update progress on a boss battle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BossBattleProgressInput"
      responses:
        "200":
          description: Updated boss battle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossBattle"

  # ── Inventory ────────────────────────────────────────────────────────────────
...
  /inventory/{itemId}/equip:
    post:
      operationId: equipItem
      tags: [inventory]
      summary: Equip a cosmetic item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"

  # ── Daily Login ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Player ─────────────────────────────────────────────────────────────────
    PlayerProfile:
      type: object
      required: [id, username, playerId, level, xp, xpToNextLevel, hp, maxHp, coins, streak, rank, rankIndex, totalXp, onboardingComplete, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        playerId:
          type: string
          description: Unique display ID like QPX-1042
        avatarUrl:
          type: ["string", "null"]
        level:
          type: integer
        xp:
          type: integer
          description: Current XP in this level
        xpToNextLevel:
          type: integer
        totalXp:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        coins:
          type: integer
        streak:
          type: integer
          description: Current daily streak in days
        longestStreak:
          type: integer
        rank:
          type: string
          description: "Current rank name: Recruit, Cadet, Trainee, Warrior..."
        rankIndex:
          type: integer
          description: 0-based rank index (0=Recruit ... 9=Titan)
        title:
          type: ["string", "null"]
          description: Equipped legendary title
        onboardingComplete:
          type: boolean
        primaryGoal:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
    PlayerProfileUpdate:
      type: object
      properties:
        username:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        title:
          type: string
    OnboardingInput:
      type: object
      required: [username, age, height, weight, primaryGoal, topPriorities, skillsToLearn, dailyHours, biggestWeakness, habitsToBuild, desiredIdentity]
      properties:
        username:
          type: string
        age:
          type: integer
        height:
          type: string
        weight:
          type: string
        primaryGoal:
          type: string
        topPriorities:
          type: array
          items:
            type: string
        skillsToLearn:
          type: array
          items:
            type: string
        dailyHours:
          type: number
        biggestWeakness:
          type: string
        habitsToLearn:
          type: array
          items:
            type: string
        habitsToLearn_extra:
          type: ["string", "null"]
          description: unused - keep for compat
        habitsToBuild:
          type: array
          items:
            type: string
        desiredIdentity:
          type: string
    XpAwardInput:
      type: object
      required: [amount, reason]
      properties:
        amount:
          type: integer
        reason:
          type: string
    XpAwardResult:
      type: object
      required: [player, xpGained, leveledUp, newLevel, newRank, rankPromoted]
      properties:
        player:
          $ref: "#/components/schemas/PlayerProfile"
        xpGained:
          type: integer
        leveledUp:
          type: boolean
        newLevel:
          type: integer
        newRank:
          type: string
        rankPromoted:
          type: boolean

    # ── Quests ─────────────────────────────────────────────────────────────────
...
    JournalEntry:
      type: object
      required: [id, title, content, mood, energyLevel, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        wins:
          type: ["string", "null"]
        lessons:
          type: ["string", "null"]
        goals:
          type: ["string", "null"]
        mood:
          type: string
          description: "excellent | good | neutral | low | terrible"
        energyLevel:
          type: integer
          description: 1-10
        createdAt:
          type: string
        updatedAt:
          type: string
    JournalEntryInput:
      type: object
      required: [title, content, mood, energyLevel]
      properties:
        title:
          type: string
        content:
          type: string
        wins:
          type: string
        lessons:
          type: string
        goals:
          type: string
        mood:
          type: string
        energyLevel:
          type: integer
    JournalEntryUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        wins:
          type: string
        lessons:
          type: string
        goals:
          type: string
        mood:
          type: string
        energyLevel:
          type: integer

    # ── Stats ──────────────────────────────────────────────────────────────────
...
    BossBattle:
      type: object
      required: [id, title, description, targetValue, currentProgress, xpReward, coinReward, status, expiresAt, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        targetValue:
          type: integer
        currentProgress:
          type: integer
        xpReward:
          type: integer
        coinReward:
          type: integer
        badgeReward:
          type: ["string", "null"]
        status:
          type: string
          description: "active | completed | failed | upcoming"
        expiresAt:
          type: string
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    BossBattleProgressInput:
      type: object
      required: [increment]
      properties:
        increment:
          type: integer

    # ── Inventory ──────────────────────────────────────────────────────────────
    Inventory:
      type: object
      required: [items, equippedOutfit, equippedEffect, equippedFrame]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        equippedOutfit:
          type: ["string", "null"]
        equippedEffect:
          type: ["string", "null"]
        equippedFrame:
          type: ["string", "null"]
    InventoryItem:
      type: object
      required: [id, name, type, rarity, unlocked, equipped]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          description: "outfit | effect | frame | badge | title"
        rarity:
          type: string
          description: "common | rare | epic | legendary"
        unlocked:
          type: boolean
        equipped:
          type: boolean
        requiredRank:
          type: ["string", "null"]
        requiredLevel:
          type: ["integer", "null"]

    # ── Daily Login ────────────────────────────────────────────────────────────
...