# lib/api-zod/src/generated/types/inventory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QPX Quantum Progress Experience API
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItem } from './inventoryItem';

export interface Inventory {
  items: InventoryItem[];
  /** @nullable */
  equippedOutfit: string | null;
  /** @nullable */
  equippedEffect: string | null;
  /** @nullable */
  equippedFrame: string | null;
}