# lib/api-zod/src/generated/types/journalEntryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QPX Quantum Progress Experience API
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntryUpdate {
  title?: string;
  content?: string;
  wins?: string;
  lessons?: string;
  goals?: string;
  mood?: string;
  energyLevel?: number;
}